correction D1

+++ fibo_iter/fibo_iter.c
@@ -4,7 +4,7 @@ unsigned long fibo_iter(unsigned long n)
     unsigned long f2 = 1;
     unsigned long temp;
 
-    for(int i = 0; i<n; i++)
+    for (unsigned long i = 0; i < n; i++)
     {
         temp = f1 + f2;
         f1 = f2;
